+++ protocols/request-response/tests/ping.rs
@@ -65,6 +65,7 @@ async fn is_response_outbound() {
             peer,
             request_id: req_id,
             error: _error,
+            ..
         } => {
             assert_eq!(&offline_peer, &peer);
             assert_eq!(req_id, request_id1);
@@ -116,6 +117,7 @@ async fn ping_protocol() {
                         request_response::Message::Request {
                             request, channel, ..
                         },
+                    ..
                 }) => {
                     assert_eq!(&request, &expected_ping);
                     assert_eq!(&peer, &peer2_id);
@@ -157,6 +159,7 @@ async fn ping_protocol() {
                             request_id,
                             response,
                         },
+                    ..
                 } => {
                     count += 1;
                     assert_eq!(&response, &expected_pong);
@@ -205,7 +208,8 @@ async fn emits_inbound_connection_closed_failure() {
             event = swarm1.select_next_some() => match event {
                 SwarmEvent::Behaviour(request_response::Event::Message {
                     peer,
-                    message: request_response::Message::Request { request, channel, .. }
+                    message: request_response::Message::Request { request, channel, .. },
+                    ..
                 }) => {
                     assert_eq!(&request, &ping);
                     assert_eq!(&peer, &peer2_id);
@@ -270,7 +274,8 @@ async fn emits_inbound_connection_closed_if_channel_is_dropped() {
             event = swarm1.select_next_some() => {
                 if let SwarmEvent::Behaviour(request_response::Event::Message {
                     peer,
-                    message: request_response::Message::Request { request, channel, .. }
+                    message: request_response::Message::Request { request, channel, .. },
+                    ..
                 }) = event {
                     assert_eq!(&request, &ping);
                     assert_eq!(&peer, &peer2_id);