+++ libp2p/CHANGELOG.md
@@ -3,6 +3,9 @@
 - Deprecate `void` crate.
   See [PR 5676](https://github.com/libp2p/rust-libp2p/pull/5676).
 
+- Expose swarm builder phase errors.
+  See [PR 5726](https://github.com/libp2p/rust-libp2p/pull/5726).
+
 ## 0.54.1
 
 - Update individual crates.