+++ protocols/autonat/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.13.2
+
+- Update to `libp2p-request-response` `v0.28.0`.
+
 ## 0.13.1
 
 - Verify that an incoming AutoNAT dial comes from a connected peer. See [PR 5597](https://github.com/libp2p/rust-libp2p/pull/5597).